Make the session data store pluggable; split JDBC into its own peer store
Turn SessionDataStore into the abstract supertype that every caller uses,
and extract the existing relational implementation into JDBCSessionDataStore
so JDBC and any alternative store (e.g. a Redis store) sit at the same
pluggable level. getInstance() selects the configured store
(JDBCPersistenceManager.SessionDataPersist.SessionStoreImplType; default
JDBC) from a registry of OSGi-registered SessionDataStore services.

No caller changes: SessionDataStore.getInstance() and every public method
keep their signatures, and all call sites — including the persist/cleanup
worker tasks and the caches — are untouched. The relational store keeps its
schema, async machinery and cleanup services verbatim (moved into
JDBCSessionDataStore). Core references only SessionDataStore and
JDBCSessionDataStore; an alternative store extends SessionDataStore and is
discovered via OSGi, so core has no dependency on it.

Store selection is fail-closed: JDBC is used only when the property is absent
or set to JDBC. When a non-JDBC store is configured but not yet registered,
resolution throws instead of falling back to JDBC (which would split session
data across two stores); the result is not cached, so once the store's OSGi
service binds the next call resolves. Store names are normalized
case-insensitively for registration/removal/lookup, and unbind removes a
mapping only when it still points to the unbound service instance.
